rfctr: consolidate pptx.opc.serialized modules
Several of these classes are going away and we don't need this many
modules to take care of the coherent serializing and deserializing
responsibilities.
